backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.7
pytesseract==0.3.10
numpy==1.26.3
Pillow==10.2.0
sympy==1.12
PyMuPDF==1.23.21
pydantic==2.5.3
python-dotenv==1.0.0